編集用画面を別ファイルに分割
