+++ app/src/main/java/si/f5/mob/photoapp/photoedit/PhotoEditViewModel.kt
@@ -1,0 +1,4 @@
+package si.f5.mob.photoapp.photoedit
+
+class PhotoEditViewModel {
+}
